+++ website_local_content/tests/__init__.py
@@ -1,5 +1,6 @@
 from . import (
     test_activity_axis,
+    test_detail_like_rating,
     test_item_start,
     test_local_content_controller,
     test_local_content_models,